+++ src/Neuron1.java
@@ -104,34 +104,26 @@ public final class Neuron1 implements Neuron {
     // These are the kernel methods that need to be implemented.
     @Override
     public int size() {
-        // TODO Auto-generated method stub
-        throw new UnsupportedOperationException("Unimplemented method 'size'");
+        return this.w.length();
     }
 
     @Override
     public double weight(int i) {
-        // TODO Auto-generated method stub
-        throw new UnsupportedOperationException(
-                "Unimplemented method 'weight'");
+        return this.w.entry(i);
     }
 
     @Override
     public void setWeights(int i, double w) {
-        // TODO Auto-generated method stub
-        throw new UnsupportedOperationException(
-                "Unimplemented method 'setWeights'");
+        this.w.replaceEntry(i, w);
     }
 
     @Override
     public double bias() {
-        // TODO Auto-generated method stub
-        throw new UnsupportedOperationException("Unimplemented method 'bias'");
+        return this.b;
     }
 
     @Override
     public void setBias(double b) {
-        // TODO Auto-generated method stub
-        throw new UnsupportedOperationException(
-                "Unimplemented method 'setBias'");
+        this.b = b;
     }
 }